Added authentication to Journal Endpoints

+++ src/main/java/xyz/znet/journalapp/service/JournalEntryService.java
@@ -28,18 +28,14 @@ public class JournalEntryService {
             journalEntry.setDate(LocalDateTime.now());
             JournalEntry saved= journalEntryRepository.save(journalEntry);
             user.getJournalEntries().add(saved);
-            userService.saveEntry(user);
+            userService.saveUser(user);
         } catch (Exception e) {
             log.error("Exception:", e);
             throw new RuntimeException("Exception:", e);
         }
 
     }
-    public void saveEntry(JournalEntry journalEntry) {
 
-      journalEntryRepository.save(journalEntry);
-
-    }
     public List<JournalEntry> getAll() {
         return journalEntryRepository.findAll();
     }
@@ -47,10 +43,22 @@ public class JournalEntryService {
     public Optional<JournalEntry> findById(ObjectId id) {
         return journalEntryRepository.findById(id);
     }
-    public void deleteById(ObjectId id, String userName) {
-        User user = userService.findByUsername(userName);
-        user.getJournalEntries().removeIf(journalEntry -> journalEntry.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName) {
+        boolean removed = false;
+        try {
+            User user = userService.findByUsername(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if (removed) {
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        } catch (Exception e) {
+            log.error("Error ",e);
+            throw new RuntimeException("An error occurred while deleting the entry.", e);
+        }
+        return removed;
     }
+
+
 }
